fourth commit

+++ src/main/java/com/example/carrentalproject/repository/ClientRepository.java
@@ -3,5 +3,5 @@ package com.example.carrentalproject.repository;
 import com.example.carrentalproject.model.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepository extends JpaRepository<Client, Long> {
+public interface ClientRepository extends JpaRepository<Client, Integer> {
 }
